feat: :fire: upgrade to allow fixable code
Upgraded the `no-moment` rule to allow fixable code, which enables automatic fixes of code that uses the Moment.js library.  Refactor the code in the `no-moment` rule to make it more robust and efficient.

Upgraded the `no-moment` rule to allow automatic fixes of code that uses the Moment.js library. This may require updates to existing code that uses this rule.